# gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Goa API
    version: 0.0.1
servers:
    - url: http://localhost:80
      description: Default server for committee
paths:
    /committees:
        post:
            tags:
                - committee
            summary: create-committee committee
            description: Create Committee
            operationId: committee#create-committee
            parameters:
                - name: v
                  in: query
                  description: Version of the API
                  allowEmptyValue: true
                  schema:
                    type: string
                    description: Version of the API
                    example: "1"
                    enum:
                        - "1"
                  example: "1"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateCommitteeRequestBody'
                        example:
                            business_email_required: false
                            category: Technical Steering Committee
                            description: Main technical oversight committee for the project
                            enable_voting: true
                            is_audit_enabled: false
                            name: Technical Steering Committee
                            parent_committee_id: 90b147f2-7cdd-157a-a2f4-9d4a567123fc
                            public: true
                            public_name: TSC Committee Calendar
                            sso_group_enabled: true
                            sso_group_name: cncf-technical-steering
                            status: active
                            website: https://committee.example.org
                            writers:
                                - manager_user_id1
                                - manager_user_id2
            responses:
                "201":
                    description: Created response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Committee'
                            example:
                                business_email_required: false
                                category: Technical Steering Committee
                                description: Main technical oversight committee for the project
                                enable_voting: true
                                id: 52ec9e74-e8d3-40d9-953c-bc2d2c6ae516
                                is_audit_enabled: false
                                name: Technical Steering Committee
                                parent_committee_id: 90b147f2-7cdd-157a-a2f4-9d4a567123fc
                                project_id: a0956000001FwZVAA0
                                public: true
                                public_name: TSC Committee Calendar
                                sso_group_enabled: true
                                sso_group_name: cncf-technical-steering
                                status: active
                                website: https://committee.example.org
                                writers:
                                    - manager_user_id1
                                    - manager_user_id2
                "400":
                    description: 'BadRequest: Bad request'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BadRequestError'
                            example:
                                message: The request was invalid.
                "409":
                    description: 'Conflict: Conflict'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ConflictError'
                            example:
                                message: The resource already exists.
                "500":
                    description: 'InternalServerError: Internal server error'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/InternalServerError'
                            example:
                                message: An internal server error occurred.
                "503":
                    description: 'ServiceUnavailable: Service unavailable'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ServiceUnavailableError'
                            example:
                                message: The service is unavailable.
            security:
                - jwt_header_Authorization: []
...
components:
    schemas:
        BadRequestError:
            type: object
            properties:
                message:
                    type: string
                    description: Error message
                    example: The request was invalid.
            example:
                message: The request was invalid.
            required:
                - message
        Committee:
            type: object
            properties:
                business_email_required:
                    type: boolean
                    description: Whether business email is required for committee members
                    default: false
                    example: false
                category:
                    type: string
                    description: The category of the committee
                    example: Technical Steering Committee
                    enum:
                        - Legal Committee
                        - Finance Committee
                        - Special Interest Group
                        - Board
                        - Technical Oversight Committee/Technical Advisory Committee
                        - Technical Steering Committee
                        - Marketing Oversight Committee/Marketing Advisory Committee
                        - Marketing Committee/Sub Committee
                        - Code of Conduct
                        - Product Security
                        - Technical Mailing List
                        - Marketing Mailing List
                        - Working Group
                        - Committers
                        - Maintainers
                        - Ambassador
                        - Government Advisory Council
                        - Expert Group
                        - Other
                description:
                    type: string
                    description: The description of the committee
                    example: Main technical oversight committee for the project
                    maxLength: 2000
                enable_voting:
                    type: boolean
                    description: Whether voting is enabled for this committee
                    default: false
                    example: true
                id:
                    type: string
                    description: The unique identifier of the committee
                    example: 52ec9e74-e8d3-40d9-953c-bc2d2c6ae516
                    format: uuid
                is_audit_enabled:
                    type: boolean
                    description: Whether audit logging is enabled for this committee
                    default: false
                    example: false
                name:
                    type: string
                    description: The name of the committee
                    example: Technical Steering Committee
                    maxLength: 100
                parent_committee_id:
                    type: string
                    description: The ID of the parent committee, should be empty if there is none
                    example: 90b147f2-7cdd-157a-a2f4-9d4a567123fc
                    format: uuid
                project_id:
                    type: string
                    description: The project identifier this committee belongs to
                    example: a0956000001FwZVAA0
                public:
                    type: boolean
                    description: Whether the committee is publicly visible
                    default: false
                    example: true
                public_name:
                    type: string
                    description: The public display name of the committee
                    example: TSC Committee Calendar
                sso_group_enabled:
                    type: boolean
                    description: Whether SSO group integration is enabled
                    default: false
                    example: true
                sso_group_name:
                    type: string
                    description: The name of the SSO group
                    example: cncf-technical-steering
                status:
                    type: string
                    description: The current status of the committee
                    default: active
                    example: active
                    enum:
                        - active
                        - inactive
                        - archived
                website:
                    type: string
                    description: The website URL of the committee
                    example: https://committee.example.org
                    format: uri
                    pattern: ^(https?://)?[^\s/$.?#].[^\s]*$
                writers:
                    type: array
                    items:
                        type: string
                        example: Voluptatum voluptas perferendis officiis similique accusantium sed.
                    description: Manager user IDs who can edit/modify this committee
                    example:
                        - manager_user_id1
                        - manager_user_id2
            description: A representation of LFX committee
            example:
                business_email_required: false
                category: Technical Steering Committee
                description: Main technical oversight committee for the project
                enable_voting: true
                id: 52ec9e74-e8d3-40d9-953c-bc2d2c6ae516
                is_audit_enabled: false
                name: Technical Steering Committee
                parent_committee_id: 90b147f2-7cdd-157a-a2f4-9d4a567123fc
                project_id: a0956000001FwZVAA0
                public: true
                public_name: TSC Committee Calendar
                sso_group_enabled: true
                sso_group_name: cncf-technical-steering
                status: active
                website: https://committee.example.org
                writers:
                    - manager_user_id1
                    - manager_user_id2
        ConflictError:
            type: object
            properties:
                message:
                    type: string
                    description: Error message
                    example: The resource already exists.
            example:
                message: The resource already exists.
            required:
                - message
        CreateCommitteeRequestBody:
            type: object
            properties:
                business_email_required:
                    type: boolean
                    description: Whether business email is required for committee members
                    default: false
                    example: false
                category:
                    type: string
                    description: The category of the committee
                    example: Technical Steering Committee
                    enum:
                        - Legal Committee
                        - Finance Committee
                        - Special Interest Group
                        - Board
                        - Technical Oversight Committee/Technical Advisory Committee
                        - Technical Steering Committee
                        - Marketing Oversight Committee/Marketing Advisory Committee
                        - Marketing Committee/Sub Committee
                        - Code of Conduct
                        - Product Security
                        - Technical Mailing List
                        - Marketing Mailing List
                        - Working Group
                        - Committers
                        - Maintainers
                        - Ambassador
                        - Government Advisory Council
                        - Expert Group
                        - Other
                description:
                    type: string
                    description: The description of the committee
                    example: Main technical oversight committee for the project
                    maxLength: 2000
                enable_voting:
                    type: boolean
                    description: Whether voting is enabled for this committee
                    default: false
                    example: true
                is_audit_enabled:
                    type: boolean
                    description: Whether audit logging is enabled for this committee
                    default: false
                    example: false
                name:
                    type: string
                    description: The name of the committee
                    example: Technical Steering Committee
                    maxLength: 100
                parent_committee_id:
                    type: string
                    description: The ID of the parent committee, should be empty if there is none
                    example: 90b147f2-7cdd-157a-a2f4-9d4a567123fc
                    format: uuid
                public:
                    type: boolean
                    description: Whether the committee is publicly visible
                    default: false
                    example: true
                public_name:
                    type: string
                    description: The public display name of the committee
                    example: TSC Committee Calendar
                sso_group_enabled:
                    type: boolean
                    description: Whether SSO group integration is enabled
                    default: false
                    example: true
                sso_group_name:
                    type: string
                    description: The name of the SSO group
                    example: cncf-technical-steering
                status:
                    type: string
                    description: The current status of the committee
                    default: active
                    example: active
                    enum:
                        - active
                        - inactive
                        - archived
                website:
                    type: string
                    description: The website URL of the committee
                    example: https://committee.example.org
                    format: uri
                    pattern: ^(https?://)?[^\s/$.?#].[^\s]*$
                writers:
                    type: array
                    items:
                        type: string
                        example: Fugiat rem similique perspiciatis saepe occaecati illo.
                    description: Manager user IDs who can edit/modify this committee
                    example:
                        - manager_user_id1
                        - manager_user_id2
            example:
                business_email_required: false
                category: Technical Steering Committee
                description: Main technical oversight committee for the project
                enable_voting: true
                is_audit_enabled: false
                name: Technical Steering Committee
                parent_committee_id: 90b147f2-7cdd-157a-a2f4-9d4a567123fc
                public: true
                public_name: TSC Committee Calendar
                sso_group_enabled: true
                sso_group_name: cncf-technical-steering
                status: active
                website: https://committee.example.org
                writers:
                    - manager_user_id1
                    - manager_user_id2
            required:
                - name
                - category
        InternalServerError:
            type: object
            properties:
                message:
                    type: string
                    description: Error message
                    example: An internal server error occurred.
            example:
                message: An internal server error occurred.
            required:
                - message
...
    securitySchemes:
        jwt_header_Authorization:
            type: http
            description: Heimdall authorization
            scheme: bearer
...